Backend/video_ai_demo/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
httpx==0.26.0
python-dotenv==1.0.0
jsonschema==4.21.1
aiosqlite==0.19.0
python-multipart==0.0.9
scenedetect[opencv]==0.6.3
opencv-python==4.9.0.80
numpy==1.26.3

# API认证和安全
PyJWT==2.8.0
email-validator==2.3.0

# 测试工具
requests==2.31.0

